renamed variables / files to follow example
