+++ src/identity-manager/SimpleCredentialMetaData.ts
