+++ src/identity-manager/VerfiableCred.ts
@@ -1,7 +1,7 @@
 import { IdentityManager } from './IdentityManager';
 import typeormConfig from './agents-ormconfig';
 import { ConnectionOptions, createConnection } from 'typeorm';
-import metadata from './metadata'
+import metadata from './SimpleCredentialMetaData'
 import { SignedCredential } from 'jolocom-lib/js/credentials/signedCredential/signedCredential';
 import fs from "fs";
 
@@ -16,28 +16,18 @@ async function init() {
 
     //Create/load first agent
     const password = readlineSync.question('Enter your password to either load or create new DID:').toString();
-    const agent = await identityMgr.loadBasedOnPass(password);
-    console.log(agent.identityWallet.did)
+    const alice = await identityMgr.loadBasedOnPass(password);
+    console.log(alice.identityWallet.did)
 
     //Create/load another agent
     const password2 = readlineSync.question('Enter another password create/load second DID:').toString();
-    const agent2 = await identityMgr.loadBasedOnPass(password2);
-    console.log(agent2.identityWallet.did)
+    const bob = await identityMgr.loadBasedOnPass(password2);
+    console.log(bob.identityWallet.did)
 
     identityMgr.logger.info("About to verify metadata between our 2 loaded/created agents")
 
-    //Const signedcredential based on certain metadata
-    const agentCredAboutagent2 = await agent.signedCredential({
-        metadata: metadata,
-        subject: agent2.identityWallet.did,
-        claim: {
-          age: 25,
-          name: 'Bob',
-        },
-      })
-
-    //Create offer token
-    const agentCredOffer = await agent.credOfferToken({
+    // Alice creates the offer to Bob to sign a simple credential
+    const aliceCredOffer = await alice.credOfferToken({
         callbackURL: 'https://example.com/issuance',
         offeredCredentials: [
         {
@@ -46,30 +36,45 @@ async function init() {
         ],
     })    
 
-    //Agent 2 proccesses offer token
-    const agent2Interaction = await agent2.processJWT(agentCredOffer.encode())
-    //Agent 2 creates resonses token
-    const agent2CredSelection = await agent2Interaction.createCredentialOfferResponseToken(
+    // Bob receives and processes the offered token, to identify the relevant Interaction 
+    const bobCredExchangeInteraction = await bob.processJWT(aliceCredOffer.encode())
+    // Bob then creates a response token
+    const bobCredExchangeResponse = await bobCredExchangeInteraction.createCredentialOfferResponseToken(
       [{ type: 'SimpleExampleCredential' }],
     );
-    agent2.processJWT(agent2CredSelection.encode());
+    // Note that all agents need to also process the tokens they generate so that their interaction manager has seen all messages
+    bob.processJWT(bobCredExchangeResponse.encode());
 
-    //Issuance and recieving
-    const agentInteraction = await agent.processJWT(agent2CredSelection.encode())
-    const agentIssuance = await agentInteraction.createCredentialReceiveToken(
-      [agentCredAboutagent2],
+    // Alice receives the token response from Bob, finds the interaction + then creates the VC to share
+    const aliceCredExchangeInteraction = await alice.processJWT(bobCredExchangeResponse.encode())
+
+    // Create the VC that then will be issued by Alice to Bob, so that Bob can then prove that Alice attested to this credential about him. 
+    const aliceAboutBobVC = await alice.signedCredential({
+      metadata: metadata,
+      subject: bob.identityWallet.did,
+      claim: {
+        age: 25,
+        name: 'Bob',
+      },
+    });
+
+    // Create the token wrapping the VC
+    const aliceCredIssuance = await aliceCredExchangeInteraction.createCredentialReceiveToken(
+      [aliceAboutBobVC],
     );
+    // Alice processes her own generated token also
+    await alice.processJWT(aliceCredIssuance.encode());
 
-    await agent.processJWT(agentIssuance.encode());
+    // Token with signed VC is sent and received by Bob
+    // Note: should be same as interaction above....check!
+    const bobCredExchangeInteraction2 = await bob.processJWT(aliceCredIssuance.encode());
 
-    const agent2Receives = await agent2.processJWT(agentIssuance.encode());
-
-    const state = (agent2Receives.getSummary() as any).state
+    const state = (bobCredExchangeInteraction2.getSummary() as any).state
 
     if (state.credentialsAllValid) {
         identityMgr.logger.info("Issued credential interaction is valid!");
         await Promise.all(
-          state.issued.map((VC: SignedCredential) =>  agent2.storage.store.verifiableCredential(VC)),
+          state.issued.map((VC: SignedCredential) =>  bob.storage.store.verifiableCredential(VC)),
         )
         identityMgr.logger.info("Saving verfied interaction^^");
         await fs.appendFileSync("VerfiedCredential.json", JSON.stringify(state)+"\r\n");
